feat: add error handling and empty data source messages for visit registration and templates
