+++ apps/frontend/src/components/template/FeedbackTemplatesPage.tsx
@@ -21,6 +21,13 @@ export default function FeedbackTemplatesPage() {
             templateGroup={templateGroup}
             itemCountLabel={itemCountLabel}
             isArchived={false}
+            emptyDataSourceMessage={
+              <div>
+                <strong>Warning</strong> Feedback has no templates. Users will
+                not be able to fill out feedback form. Please create a new
+                template and mark it as active.
+              </div>
+            }
           />
           <TableComponent
             templateGroup={templateGroup}
